refactor: rename health to symptom
- Rename pages/health/ to pages/symptom/
- Rename services/health.ts to services/symptom.ts
- Rename constants/health.ts to constants/symptom.ts
- Rename HealthRecord type to SymptomRecord
- Rename database collection health_records to symptom_records

The naming better reflects the parallel relationship with other
record types: symptom, meal, stool, medication, exam.

Closes #18
