+++ src/app.config.ts
@@ -1,5 +1,5 @@
 export default defineAppConfig({
-  pages: ["pages/index/index", "pages/health/index/index", "pages/health/add/index"],
+  pages: ["pages/index/index", "pages/symptom/index/index", "pages/symptom/add/index"],
   window: {
     backgroundTextStyle: "light",
     navigationBarBackgroundColor: "#fff",
