+++ src/constants/symptom.ts
