+++ src/pages/symptom/add/index.config.ts
