+++ src/pages/symptom/add/index.css
