+++ src/pages/symptom/index/index.config.ts
