+++ src/pages/symptom/index/index.css
